requirement.txt
langchain == 1.2.0
python_dotenv ==1.2.1
langchain_community == 0.4.1
langchain_groq == 1.1.1
streamlit == 1.52.2